Allow entering some of the mode and frequency commands
N1MM has the option to enter the frequency or the mode in the callsign
field and press return. That will make the rig jump to that mode or
frequency. A similar, but most likely simpler - I did not experiement
on the nitty gritty details of N1MM, is added by this change.