Fix crash opening windows on systems with zombie processes
When opening new windows (band map, log, etc), we currently iterate over
all running processes on the system to ensure that window is not
currently open. However, if there are any zombie processes (unrelated to
us), `proc.cmdline()` will throw an exception.

This commit simply catches and continues when that exception is thrown.
It will also catch NoSuchProcess in the event that a process closed
between grabbing the list and that process in the iteration.